pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bsky-pulse"
version = "0.1.0"
description = "Live sentiment analysis pipeline for Bluesky (AT Protocol) firehose data"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "websockets>=12.0",
    "motor>=3.3",
    "pymongo>=4.6",
    "pandas>=2.2",
    "wordcloud>=1.9",
    "atproto>=0.0.46",
    "certifi>=2024.0",
    "python-dotenv>=1.0",
    "vaderSentiment>=3.3",
    "matplotlib>=3.8",
    "notebook>=7.0",
    "ipywidgets>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.setuptools.packages.find]
where = ["src"]
